adding user authentication

+++ graphql/resolvers/booking.js
@@ -5,7 +5,8 @@ const { transformEvent,
 
 
 module.exports = { 
-    bookings: async () => {
+    bookings: async (args, req) => {
+        if(!req.isAuth) throw new Error('Unauthenticated!!');
         try {
             const bookings = await Booking.find();
             return bookings.map(booking => {
@@ -15,22 +16,27 @@ module.exports = {
             throw err;
         }
     },
-    bookEvent: async args => {
+
+    bookEvent: async (args,req ) => {
+        if(!req.isAuth) throw new Error('Unauthenticated!!');
+
         const fetchedEvent = await Event.findOne({_id: args.eventId});
         const booking = new Booking(
-            {user: "5d8265e6caf9890431e8a657", event: fetchedEvent}
+            {user: req.userId, event: fetchedEvent}
         );
 
         const result = await booking.save();
         return transfromBooking(result);
     },
-    cancelBooking: async args => {
+
+    cancelBooking: async (args,req) => {
+        if(!req.isAuth) throw new Error('Unauthenticated!!');
         try {
             const booking = await Booking
                 .findById(args.bookingId)
                 .populate('event');
             const event = transformEvent(booking.event); //booking._doc.event is also working.
-
+            
             await Booking.deleteOne({_id: args.bookingId});
             return event;
         } catch (err) {
